requirements.txt
aws-cdk-lib==2.149.0
constructs>=10.0.0,<11.0.0
kaggle==1.5.12
pandas==2.2.2
numpy==1.23.5
pillow==10.4.0
boto3==1.34.152
opensearch-py==2.3.1
retrying==1.3.4
streamlit==1.37.1
langchain==0.2.14
langchain-community==0.2.12
streamlit-js-eval==0.1.7
python-dotenv==1.0.1